+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Backing up a calendar that has the same name as the default calendar
 - Added additional backoff-retry to all OneDrive queries.
+- Users with `null` userType values are no longer excluded from user queries.
 
 ### Known Issues